refactor: create more components from page
