+++ src/entities/listItem.ts
@@ -1,0 +1,7 @@
+export interface ListItemData {
+  [index: string]: string | number
+
+  name: string
+  quantity: number
+  price: number
+}
